third_party/blink/web_tests/external/wpt/css/css-fonts/parsing/font-style-computed-expected.txt

This is a testharness.js-based test.
[FAIL] Property font-style value 'oblique'
  assert_equals: expected "oblique" but got "italic"
[FAIL] Property font-style value 'oblique 0deg'
  assert_equals: expected "oblique 0deg" but got "normal"
Harness: the test ran to completion.
